+++ src/core/misc/useTweakPane/index.ts
@@ -1,50 +1,49 @@
 import { onUnmounted, onMounted } from 'vue'
 import { Pane } from 'tweakpane'
-import * as EssentialsPlugin from '@tweakpane/plugin-essentials'
 import { useRenderLoop } from '@tresjs/core'
+import { useLogger } from '../../../composables/useLogger'
 
-type TweakPane = Pane & { addBlade(blade: any): void }
+type TweakPane = Pane & { addBlade(blade: any): void, addInput(blade: any): void }
 let pane: TweakPane
-let fpsGraph: any
 
 export /**
  * Creates a TweakPane instance and returns it.
  *
- * @param {string} [selector='tres-container']
+ * @param {string} [selector='body']
  * @return {*}
  */
-const useTweakPane = (selector = 'tres-container') => {
-  if (!pane) {
-    pane = new Pane({
-      container: (document.querySelector(selector) as HTMLElement) || undefined,
-    }) as TweakPane
-    pane.registerPlugin(EssentialsPlugin)
-
-    fpsGraph = pane.addBlade({
-      view: 'fpsgraph',
-      label: 'fpsgraph',
-    })
-  }
-
-  /**
-   * Disposes the TweakPane instance.
-   *
-   */
-  function disposeTweakPane() {
-    if (pane) {
-      pane.dispose()
+  const useTweakPane = (selector = 'body') => {
+    if (!pane) {
+      pane = new Pane({
+        container: document.querySelector(selector) as HTMLElement,
+      }) as TweakPane
+      pane.element.style.position = 'absolute'
+      pane.element.style.top = '1rem'
+      pane.element.style.right = '1rem'
+      pane.element.style.zIndex = '9999'
     }
+
+    const { logWarning } = useLogger()
+
+    logWarning('useTweakPane is going to be deprecated soon and will no longer be part of this package, please start migrating to @tresjs/leches package https://github.com/Tresjs/leches or v-tweakpane https://github.com/vinayakkulkarni/v-tweakpane instead.')
+
+    /**
+     * Disposes the TweakPane instance.
+     *
+     */
+    function disposeTweakPane() {
+      if (pane) {
+        pane.dispose()
+      }
+    }
+
+    onMounted(() => {
+      const { resume } = useRenderLoop()
+      resume()
+    })
+    onUnmounted(() => {
+      disposeTweakPane()
+    })
+
+    return { pane, disposeTweakPane }
   }
-
-  onMounted(() => {
-    const { onBeforeLoop, onAfterLoop, resume } = useRenderLoop()
-    resume()
-    onBeforeLoop(() => fpsGraph.begin())
-    onAfterLoop(() => fpsGraph.end())
-  })
-  onUnmounted(() => {
-    disposeTweakPane()
-  })
-
-  return { pane, fpsGraph, disposeTweakPane }
-}